GUI: render canonical display names in LLM Chat model dropdown
The Active Models row used getModelDisplayName() to turn shadowed canonical
ids (user.*, extra.*, builtin.*) into "Foo (registered)" form, while the
LLM Chat dropdown showed the raw id. Extract getModelDisplayName and
CANONICAL_PREFIXES into a shared util so both call sites render the same
string. The dropdown search also matches against the display string.
